% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{add_model_content}
\alias{add_model_content}
\title{Add model content}
\usage{
add_model_content(session, file, model, role = NULL, exact = TRUE, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{file}{path to file}

\item{model}{\code{MMmodel} object, model ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{role}{file role, such as "scoreResource" and "score"}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{pass to \code{sasctl::vPOST()} function}
}
\value{
a \code{MMcontent} class list
}
\description{
Add model content
}
\examples{

\dontrun{
my_model <- get_model(sess, "MyModel")

myContent <- add_model_content(sess, file = "my_fancy_file.R", model = my_model)
myContent
}

}
