% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{add_model_version}
\alias{add_model_version}
\title{Add model version}
\usage{
add_model_version(session, model, exact = TRUE, minor = FALSE, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{model}{\code{MMmodel} object, project ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{minor}{if \code{TRUE}, new minor version is created, by default a major version is created.}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of projects
}
\description{
Add model version
}
\examples{

\dontrun{
my_model <- get_model(sess, model = "MyModel")

nvmodel <- add_model_version(sess,  my_model)

nvmodel
}

}
