% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{calculateROCStat}
\alias{calculateROCStat}
\title{Write dmcas_roc Json}
\usage{
calculateROCStat(
  targetName,
  targetPredicted,
  validadedf = NULL,
  traindf = NULL,
  testdf = NULL,
  targetEventValue = 1,
  label.ordering = c(0, 1),
  path = NULL,
  noFile = FALSE
)
}
\arguments{
\item{targetName}{target variable column name (actuals)}

\item{targetPredicted}{target variable predicted probability column name}

\item{validadedf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{traindf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{testdf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{targetEventValue}{target class name for ROC reference, if model is nominal, all other class will be counted as "not target"}

\item{label.ordering}{The default ordering (cf.details) of the classes can be changed by supplying a vector containing the negative and the positive class label. See \code{\link[ROCR:prediction]{ROCR::prediction()}}}

\item{path}{directory where the JSON file is written; required when
\code{noFile = FALSE}}

\item{noFile}{if you don't want to write to a file, only the output}
}
\value{
\itemize{
\item \code{list} that reflects the 'dmcas_roc.json'
\item 'dmcas_roc.json' file written to \code{path}
}
}
\description{
Calculates the ROC curve from user data and writes it to a JSON file for
importing into the common model repository. Binary response only.
}
\examples{

df <- data.frame(label = sample(c(1,0), 6000, replace = TRUE),
                 prob = runif(6000),
                 partition = rep_len(1:3, 6000)) ## partition will be ignored since it is 3rd column
               
calculateROCStat(targetName = "label",
                 targetPredicted = "prob",
                 df[df$partition == 1, ],
                 df[df$partition == 2, ],
                 df[df$partition == 3, ],
                 noFile = TRUE)
                                    
                                    
 
}
