% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen.glm}
\alias{codegen.glm}
\title{Code generator for General Linear models, specifically logistic regression}
\usage{
\method{codegen}{glm}(model, path = "scoreCode.R", rds = "model.rds", cutoff = 0.5)
}
\description{
Code generator for General Linear models, specifically logistic regression
}
