% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen}
\alias{codegen}
\alias{codegen.lm}
\alias{codegen.glm}
\alias{codegen.workflow}
\title{SAS standard score code generation Generic Function \strong{(EXPERIMENTAL)}}
\usage{
codegen(
  model,
  path,
  rds,
  libs,
  inputs,
  output_as_df = TRUE,
  add_target_name = FALSE,
  ...
)

\method{codegen}{lm}(
  model,
  path = "scoreCode.R",
  rds = "model.rds",
  libs = c(),
  inputs = NULL,
  output_as_df = TRUE,
  add_target_name = FALSE,
  ...
)

\method{codegen}{glm}(
  model,
  path = "scoreCode.R",
  rds = "model.rds",
  libs = c(),
  inputs = NULL,
  output_as_df = TRUE,
  add_target_name = FALSE,
  cutoff = 0.5,
  ...
)

\method{codegen}{workflow}(
  model,
  path = "scoreCode.R",
  rds = "model.rds",
  libs = c(),
  inputs = NULL,
  output_as_df = TRUE,
  add_target_name = FALSE,
  referenceLevel = NULL,
  ...
)
}
\arguments{
\item{model}{model object (lm, glm, tidymodels workflow, ...)}

\item{path}{file name and path to write}

\item{rds}{.rds file name to be called}

\item{libs}{vector of libraries to be added to the code. Some may be guessed from the type.}

\item{inputs}{define inputs as the passed vector instead of guessed}

\item{output_as_df}{logical; when \code{TRUE} (default) the score function returns a \code{data.frame}, otherwise a list}

\item{add_target_name}{logical; when \code{TRUE} (default) includes \verb{<<target>> = <<target>>} in the code output}

\item{...}{to be passes to individual code generators}

\item{cutoff}{classification probability cutoff}

\item{referenceLevel}{reference level for a factor target value}
}
\value{
a code string
}
\description{
\strong{EXPERIMENTAL STATE - MAY NOT WORK AS INTENDED}

Score code will only be generated successfully for supported models.
Other models and frameworks will be added in due time.
Use \code{\link[=create_scoreSample]{create_scoreSample()}} to get a structure sample

Disclaimer: The score code that is generated is designed to be a working template for an R model,
but is not guaranteed to work out of the box for scoring, publishing, or validating the model.
}
\section{Methods (by class)}{
\itemize{
\item \code{codegen(lm)}: Code generator for \code{lm} class models

\item \code{codegen(glm)}: generator for \code{glm} class models, specifically logistic regression

\item \code{codegen(workflow)}: generator for \link[tidymodels]{tidymodels} \code{workflow} class models

}}
\examples{

\dontrun{
# SAS viya doesn't play nice with variables with '.' in the names
colnames(iris) <- gsub("\\\\.", "_", colnames(iris))

# simple regression
model <- lm(Petal.Length ~ ., data = iris)

codegen(model)
} 

}
