% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{create_project}
\alias{create_project}
\title{Create a project}
\usage{
create_project(
  session,
  name,
  description = NULL,
  model_function = NULL,
  input_vars = NULL,
  output_vars = NULL,
  image = NULL,
  additional_parameters = NULL,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{name}{The name of the project}

\item{description}{The description of the project.}

\item{model_function}{The project model function of the project. Valid values: analytical, classification, cluster, forecasting, prediction, Text categorization, Text extraction, Text sentiment, Text topics, transformation}

\item{input_vars}{\code{data.frame} with the input data sample to configure the project variables.}

\item{output_vars}{\code{data.frame} with the output data sample to configure the project variables.}

\item{image}{Image URI to be used as project cover}

\item{additional_parameters}{\code{list} of parameters and lists to be added to the payload}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of projects
}
\description{
Returns a sasctl MMproject object from Model Manager
}
\examples{

\dontrun{
new_project <- create_project(sess, name = "ModelProj", 
                                    description = "My fancy project",
                                    model_function = "classification")
new_project
}

}
