% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{create_scoreSample}
\alias{create_scoreSample}
\title{Create Score Code template}
\usage{
create_scoreSample(path = NULL, openFile = TRUE)
}
\arguments{
\item{path}{directory where the example file is created; required}

\item{openFile}{automatically open file for editing}
}
\value{
nothing
}
\description{
Creates an R file in the path with an example.  The file structure are as follows:
For official documentation go to \href{https://go.documentation.sas.com/doc/en/mdlmgrcdc/v_025/mdlmgrug/n04i7s6bdu7ilgn1e350am3byuxx.htm#p0m6jg3tgih1agn1cgdcexpwgp0g}{Scoring R models documentation}
}
\details{
\itemize{
\item The file should start with a function with all the input variables which SAS Viya will use to insert data
\item Then it is followed by a comment line \verb{#output: outvar1, outvar2} which is case sensitive and must follow that
structure so SAS can receive the function output properly.
\item If you are using a previously created model, it should be read, we recommend \code{.rda} format, but could be
a \code{pmml} file or other format that suits you, just make sure that it is properly classified as scoring resource
when using inside SAS Model Manager.
\item You then can use any logic to score the model or just an arbitrary R code.
\item To pass the information back to SAS it must return a list of the variables defined at the beginning of the
script.
}
}
\examples{

\dontrun{
create_scoreSample(path = tempdir(), openFile = FALSE)
}

# SAS does not expect the following outputs necessarily
# but if you follow that structure it will play nice with other SAS Features

# EM_CLASSIFICATION - Predicted for target
# EM_EVENTPROBABILITY - Probability target=1 
# EM_PROBABILITY - Probability of Classification
# I_<<target>> - eg.: I_BAD - Into: BAD
# I_<<target>><<level>> eg.: I_BAD1 - predicted level

}
