% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logon_clients.R
\name{delete_client}
\alias{delete_client}
\title{Delete a client}
\usage{
delete_client(session, client)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{client}{\code{sasClient} object or client ID.}
}
\value{
A \code{\link[httr:response]{httr::response}} object.
}
\description{
Delete a client
}
\examples{

\dontrun{
new_client <- register_client(sess, 'my_client', 'my_s3cr3t!')
delete_client(sess, "my_client")
}

}
