% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{delete_model}
\alias{delete_model}
\title{Delete a model}
\usage{
delete_model(session, model, exact = TRUE)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{model}{\code{MMmodel} or \code{MMmodelVersion} object, model ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}
}
\value{
A \code{\link[httr:response]{httr::response}} object.
}
\description{
delete a model from Model Manager
}
\examples{

\dontrun{
my_model <- get_model(sess, "MyModel")

delete_model(sess, my_model)
}

}
