% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{delete_model_contents}
\alias{delete_model_contents}
\title{Delete a model content}
\usage{
delete_model_contents(session, model, content, exact = TRUE)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{model}{\code{MMmodel} object, model ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{content}{\code{MMmodelContentList} obtained from \code{\link[=list_model_contents]{list_model_contents()}}, if missing will delete all files will be deleted.}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}
}
\value{
A \code{\link[httr:response]{httr::response}} object.
}
\description{
delete model from Manager
}
\examples{

\dontrun{
my_model <- get_model(sess, "MyModel")

delete_model_contents(sess, my_model)
}

}
