% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project}
\usage{
delete_project(session, project, exact = TRUE)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{project}{\code{MMproject} object, project ID or project name. If name, will try to find a single project with exact name match. See \code{exact} parameter}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}
}
\value{
A \code{\link[httr:response]{httr::response}} object.
}
\description{
Delete a project and all associated models and resources
}
\examples{

\dontrun{
new_project <- create_project(sess, name = "ModelProj", 
                                    description = "My fancy project", 
                                    model_function = "classification")
                                    
delete_project(sess, new_project)
}

}
