% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{diagnosticsJson}
\alias{diagnosticsJson}
\title{Write all diagnostic Json files}
\usage{
diagnosticsJson(
  targetName,
  targetPredicted,
  validadedf = NULL,
  traindf = NULL,
  testdf = NULL,
  type = "binary",
  targetEventValue = 1,
  cutoff = 0.5,
  label.ordering = c(0, 1),
  path = NULL,
  noFile = FALSE
)
}
\arguments{
\item{targetName}{target variable name (actuals)}

\item{targetPredicted}{target variable probability column name}

\item{validadedf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{traindf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{testdf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{type}{\code{"binary"} or \code{"interval"}}

\item{targetEventValue}{if \code{type = "binary"} target class name for fit stat reference, if model is nominal, all other class will be counted as "not target"}

\item{cutoff}{cutoff to be used for calculation of miss classification for binary}

\item{label.ordering}{The default ordering (cf.details) of the classes can be changed by supplying a vector containing the negative and the positive class label. See \code{\link[ROCR:prediction]{ROCR::prediction()}}}

\item{path}{directory where the JSON files are written; required when
\code{noFile = FALSE}}

\item{noFile}{if you don't want to write to a file, only the output}
}
\value{
\itemize{
\item \code{list} of lists that reflects the 'dmcas_fitstat.json', 'dmcas_roc.json' and 'dmcas_lift.json'
\item 'dmcas_fitstat.json', 'dmcas_roc.json' and 'dmcas_lift.json' files written to \code{path}
}
}
\description{
Calculates and writes fit statistics, roc and lift for binary and fit statistics for
interval.
}
\examples{

df <- data.frame(label = sample(c(1,0), 6000, replace = TRUE),
                prob = runif(6000),
               partition = rep_len(1:3, 6000)) ## partition will be ignored since it is 3rd column
               
diagnosticsJson(df[df$partition == 1, ], 
                 df[df$partition == 2, ],
                 df[df$partition == 3, ],
                 targetName = "label",
                 targetPredicted = "prob",
                 noFile = TRUE
                 )
                                    
                                    
 
}
\seealso{
All parameters are passed to \code{\link[=calculateLiftStat]{calculateLiftStat()}}, \code{\link[=calculateLiftStat]{calculateLiftStat()}} and \code{\link[=calculateLiftStat]{calculateLiftStat()}} for matching parameters.
}
