% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{format_data_json}
\alias{format_data_json}
\title{Format Data.Frame rows to json format}
\usage{
format_data_json(df, scr = FALSE, metadata_columns = NULL)
}
\arguments{
\item{df}{data frame to be transformed in JSON format rows}

\item{scr}{boolean, if \code{TRUE} will write the new json format with metadata}

\item{metadata_columns}{columns names to be used as metadata. If scr is set to \code{FALSE}, metatada_columns is ignored}
}
\value{
a vector of JSON strings
}
\description{
Viya MAS requires a very specific json format which is the goal of this function
to create
}
\examples{

json_output <- format_data_json(mtcars)
json_output

}
