% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logon_clients.R
\name{get_client}
\alias{get_client}
\title{Get a client}
\usage{
get_client(session, client, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{client}{\code{sasClient} object or client ID.}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
\code{list} with sasctl attribute
}
\description{
Returns a single sasctl \code{MMclient} from SAS Model Manager
}
\examples{

\dontrun{
my_client <- get_client(sess, client = ModelProj)
my_client
}

}
