% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microanalyticScore.R
\name{get_masmodule}
\alias{get_masmodule}
\title{Get a MAS module/model and steps}
\usage{
get_masmodule(session, module, id = NULL, exact = TRUE, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{module}{\code{MASmodule} object, module name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{id}{module id, will replace \code{module} parameter if given}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
\code{list} with sasctl attribute
}
\description{
Returns a single module \code{MASmodule} object with module and steps information
}
\examples{

\dontrun{
my_module <- get_masmodule(sess, model = "name")
my_module
}

}
