% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{get_model}
\alias{get_model}
\title{Get a model}
\usage{
get_model(session, model, exact = TRUE, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{model}{\code{MMmodel} or \code{MMmodelVersion} object, project ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
\code{list} with sasctl attribute
}
\description{
Returns a single sasctl \code{MMmodel} from SAS Model Manager
}
\examples{

\dontrun{
my_model <- get_model(sess, model = MMmodel)

my_model
}

}
