% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{get_project}
\alias{get_project}
\title{Get a project}
\usage{
get_project(session, project, exact = TRUE, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{project}{\code{MMproject} object, project ID or project name. If name, will try to find a single project with exact name match. See \code{exact} parameter}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
\code{list} with sasctl attribute
}
\description{
Returns a single sasctl \code{MMProject} from SAS Model Manager
}
\examples{

\dontrun{
my_project <- get_project(sess, project = ModelProj)

my_project
}

}
