% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logon_clients.R
\name{list_clients}
\alias{list_clients}
\title{List Clients}
\usage{
list_clients(
  session,
  start = 1,
  count = 100,
  filter = NULL,
  exact = FALSE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{start}{the index of the first project to return}

\item{count}{The number of results per page. The default is 100.}

\item{filter}{character string of client_id name to be filtered}

\item{exact}{boolean, If the filter query should use "co" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} of \code{sasclientList} class with the list of clients
}
\description{
Returns a list of clients
}
\examples{

\dontrun{
clients <- list_clients(sess, filter = list(createdBy = "creatorUser", name = "projectName"))
clients
}

}
