% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{list_model_contents}
\alias{list_model_contents}
\title{List Models}
\usage{
list_model_contents(session, model, start = 0, limit = 20, exact = FALSE, ...)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{model}{\code{MMmodel} or \code{MMmodelVersion} object, model ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{start}{the index of the first content to return}

\item{limit}{maximum number of models to return}

\item{exact}{boolean, If the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
A \code{MMmodelContentList} list with the list of contents
}
\description{
Returns a list of models from Model Manager
}
\examples{

\dontrun{
models <- list_models(sess, filter = list(createdBy = "creatorUser", name = "modelName"))

models
}

}
