% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microanalyticScore.R
\name{list_modules}
\alias{list_modules}
\title{Get list of available models}
\usage{
list_modules(
  session,
  filters = list(),
  start = 0,
  limit = 20,
  verbose = FALSE,
  exact = FALSE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{filters}{list of of names vectors for filter parameters (createdBy, modifiedBy, name). By default it will use the \code{contains} operation. For more info visit \verb{https://developer.sas.com/apis/rest/Topics/#filters}.}

\item{start}{the index of the first module to return}

\item{limit}{maximum number of modules to return}

\item{verbose}{logical, return print API call information}

\item{exact}{boolean, If the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of models
}
\description{
Return a \code{data.frame} of metadata of available models/decisions
}
\examples{

\dontrun{
models <- list_modules(sess, filters = list(createdBy = 'myUser'))
models
}

}
