% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{list_projects}
\alias{list_projects}
\title{List Projects}
\usage{
list_projects(
  session,
  start = 0,
  limit = 10,
  filters = list(),
  exact = FALSE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{start}{the index of the first project to return}

\item{limit}{maximum number of projects to return}

\item{filters}{list of of names vectors for filter parameters (createdBy, modifiedBy, name). By default it will use the \code{contains} operation. For more info visit \verb{https://developer.sas.com/apis/rest/Topics/#filters}.}

\item{exact}{boolean, If the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of projects
}
\description{
Returns a list of projects from Model Manager
}
\examples{

\dontrun{
projects <- list_projects(sess, filter = list(createdBy = "creatorUser", name = "projectName"))
projects
}

}
