% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{list_repositories}
\alias{list_repositories}
\title{List Repositories}
\usage{
list_repositories(
  session,
  start = 0,
  limit = 10,
  filters = list(),
  exact = FALSE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{start}{the index of the first project to return}

\item{limit}{maximum number of repositories to return}

\item{filters}{list of of names vectors for filter parameters (createdBy, modifiedBy, name). By default it will use the \code{contains} operation. For more info visit \verb{https://developer.sas.com/apis/rest/Topics/#filters}.}

\item{exact}{boolean, If the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of repositories
}
\description{
Returns a list of model repositories. This is required to be able to create model projects
}
\examples{

\dontrun{
repositories <- list_repositories(sess, filter = 
                                            list(createdBy = "creatorUser",
                                                 name = "projectName")
                                                 )
repositories
}

}
