% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{refresh_session}
\alias{refresh_session}
\title{Refresh a Viya Session token}
\usage{
refresh_session(session, verbose = FALSE)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{verbose}{logical, return print API call information}
}
\value{
A \code{viya_connection} object with its access token and related
authentication fields updated from the refresh-token response.
}
\description{
Refresh a viya session object Oauth token
}
\examples{

\dontrun{
sess <- refresh_session(sess)
}

}
\seealso{
Other authentication:
\code{\link[=oauth_consul]{oauth_consul()}},
\code{\link[=session]{session()}}
}
\concept{authentication}
