% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{session}
\alias{session}
\title{Viya Session}
\usage{
session(
  hostname,
  username = NULL,
  password = NULL,
  client_id = NULL,
  client_secret = NULL,
  oauth_token = NULL,
  authinfo = NULL,
  auth_code = FALSE,
  verbose = FALSE,
  verify_ssl = TRUE,
  cacert = NULL,
  openBrowser = TRUE,
  platform = TRUE
)
}
\arguments{
\item{hostname}{string, SAS Viya url}

\item{username}{string, username for login}

\item{password}{string, username password}

\item{client_id}{string, client_id used for authentication, if left blank will use default}

\item{client_secret}{string, client_secret used for authentication, if left blank will use default}

\item{oauth_token}{string, if Oauth token is provided, a viya_connection is created}

\item{authinfo}{A \code{character} string that specifies an alternative path to a \code{.authinfo} file that is used for authentication. By default, \verb{~/.authinfo} is used on Linux and \verb{\%HOMEDRIVE\%} or \verb{\\\%HOMEPATH\%\\_authinfo} is used on Windows.}

\item{auth_code}{logical, if TRUE will open a browser with the user and request the authentication code to continue the authentication process. Viya 2022+ only.}

\item{verbose}{logical, return print API call information}

\item{verify_ssl}{boolean, verify SSL (Use it with caution)}

\item{cacert}{ca certificate list}

\item{openBrowser}{boolean, if \code{TRUE} the browser be opened automatically when \code{auth_code = TRUE}}

\item{platform}{logical, make a get call to get platform information (release, OS, siteName)}
}
\value{
\code{viya_connection} class object
}
\description{
Creates a Viya session object to be used in other calls
}
\examples{

\dontrun{
sess <- session(hostname = "http://myserver.com",
         username = "myuser",
         password = "mysecret")
}

}
\seealso{
Other authentication:
\code{\link[=oauth_consul]{oauth_consul()}},
\code{\link[=refresh_session]{refresh_session()}}
}
\concept{authentication}
