% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{update_model}
\alias{update_model}
\title{Update a model}
\usage{
update_model(
  session,
  name,
  model,
  input_vars = NULL,
  output_vars = NULL,
  additional_parameters = NULL,
  update_variables = TRUE,
  exact = TRUE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{name}{The name of the model}

\item{model}{\code{MMmodel} object, model ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{input_vars}{\code{data.frame} with the input data sample to configure the model variables.}

\item{output_vars}{\code{data.frame} with the output data sample to configure the model variables.}

\item{additional_parameters}{\code{list} of parameters and lists to be added to the payload}

\item{update_variables}{logical, TRUE, will make additional rest call to update variables using \code{\link[=update_model_variables]{update_model_variables()}},
variables can't be changed in the same endpoint as other resources}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}
}
\value{
A \code{MMmodel} object with the model information
}
\description{
Returns a sasctl MMmodel object from Model Manager
}
\examples{

\dontrun{
updated_model <- update_model(sess, model = new_model, 
                              additional_parameters = 
                               list(description = "Updated fancy description",
                              modeler = "BAD"))
updated_model
}

}
