% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{update_model_variables}
\alias{update_model_variables}
\title{Update model variables}
\usage{
update_model_variables(
  session,
  model,
  input_vars = NULL,
  output_vars = NULL,
  exact = TRUE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{model}{\code{MMmodel} object, model ID or model name. If name, will try to find a single model with exact name match. See \code{exact} parameter}

\item{input_vars}{\code{data.frame} with the input data sample to configure the model variables.}

\item{output_vars}{\code{data.frame} with the output data sample to configure the model variables.}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}
}
\value{
A \code{MMmodelVariables} variables
}
\description{
Update a SAS Model Manager model variables
}
\examples{

\dontrun{
new_variables <- update_model_variables(sess, model = new_model, 
                                            input_vars = iris[,1:4], 
                                            output_vars = iris[,5, drop = F]
                             )
new_variables
}

}
