% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{update_project}
\alias{update_project}
\title{Update a project}
\usage{
update_project(
  session,
  name,
  project,
  input_vars = NULL,
  output_vars = NULL,
  additional_parameters = NULL,
  update_variables = TRUE,
  exact = TRUE,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{name}{The name of the project}

\item{project}{\code{MMproject} object, project ID or project name. If name, will try to find a single project with exact name match. See \code{exact} parameter}

\item{input_vars}{\code{data.frame} with the input data sample to configure the project variables.}

\item{output_vars}{\code{data.frame} with the output data sample to configure the project variables.}

\item{additional_parameters}{\code{list} of parameters and lists to be added to the payload}

\item{update_variables}{logical, TRUE, will make additional rest call to update variables using \code{\link[=update_project_variables]{update_project_variables()}},
variables can't be changed in the same endpoint as other resources}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of projects
}
\description{
Returns a sasctl MMproject object from Model Manager
}
\examples{

\dontrun{
new_project <- create_project(sess, name = "ModelProj", 
                              description = "My fancy project",
                              model_function = "classification")
                              
updated_project <- update_project(sess, project = new_project, 
                                  additional_parameters = 
                                  list(description = "Updated fancy description",
                                       scoreInputTable = "myTable",
                                       predictionVariable = "BAD"))
updated_project
}

}
