% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_repository.R
\name{update_project_variables}
\alias{update_project_variables}
\title{Update project variables}
\usage{
update_project_variables(
  session,
  project,
  input_vars = NULL,
  output_vars = NULL,
  exact = TRUE,
  sasctl_vars,
  ...
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{project}{\code{MMproject} object, project ID or project name. If name, will try to find a single project with exact name match. See \code{exact} parameter}

\item{input_vars}{\code{data.frame} with the input data sample to configure the project variables.}

\item{output_vars}{\code{data.frame} with the output data sample to configure the project variables.}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}

\item{sasctl_vars}{\code{data.frame} from \code{MMmodel} object \code{inputVariables}, \code{outputVariables}.}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}
}
\value{
A \code{data.frame} with the list of projects
}
\description{
Returns a sasctl MMproject object from Model Manager
}
\examples{

\dontrun{
new_project <- create_project(sess, name = "ModelProj",  
                                    description = "My fancy project", 
                                    model_function = "classification")
                                    
new_variables <- update_project_variables(sess, project = new_project, 
                                          input_vars = iris[,1:4], 
                                          output_vars = iris[,5, drop = F])
                                          
new_variables
}

}
