% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{vDELETE}
\alias{vDELETE}
\title{Viya DELETE}
\usage{
vDELETE(session, path, ..., verbose = FALSE)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{path}{character, path to the GET api endpoint}

\item{...}{additional parameters to be passed to \code{httr::DELETE} such as \code{httr::add_headers}}

\item{verbose}{logical, return print API call information}
}
\value{
An \code{httr::response()} type object
}
\description{
Wrapper to make generic DELETE calls to SAS Viya
}
\examples{

\dontrun{
folders <- vGET(session,
                path = "folders/folders/")


newFolder <- vPOST(session,
                   path = "folders/folders/",
                   query = list(parentFolderUri = folders$items$parentFolderUri[1]),
                   payload = list(name = "newFolder"))


deletedFolder <- vDELETE(session,
                         path = "folders/folders/",
                         resourceID = newFolder$id)

}
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}, \code{\link[httr:PUT]{httr::PUT()}}, \code{\link[httr:DELETE]{httr::DELETE()}}

Other core API requests:
\code{\link[=vGET]{vGET()}},
\code{\link[=vHEAD]{vHEAD()}},
\code{\link[=vPOST]{vPOST()}},
\code{\link[=vPUT]{vPUT()}}
}
\concept{core API requests}
