% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{vPOST}
\alias{vPOST}
\title{Viya POST}
\usage{
vPOST(
  session,
  path,
  payload,
  ...,
  query,
  fragment,
  encode = "json",
  verbose = FALSE,
  output = "json"
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{path}{character, path to the GET api endpoint}

\item{payload}{list or json string, if it is a list, will be transformed in a json string using \code{jsonlite::toJSON}}

\item{...}{additional parameters to be passed to \code{httr::POST} such as \code{httr::add_headers}}

\item{query}{list, additional URL query parameters}

\item{fragment}{string, additional URL fragment parameter \code{url.com?query#fragment}}

\item{encode}{payload encoding type, to be passed to \code{httr::POST}.}

\item{verbose}{logical, return print API call information}

\item{output}{string, if \code{json} will return \code{httr::fromJSON(httr::content(response, as = "text"))} with etag information parsed,
else if \code{response} will return a \code{httr::response()} object, if \code{text} will return \code{content(response, as = "text")}}
}
\value{
list if \code{output = "json"} default. \code{httr::response()} if \code{output = response}.
}
\description{
Wrapper to make generic POST calls to SAS Viya
}
\examples{

\dontrun{
folders <- vGET(session,
                path = "folders/folders/")


newFolder <- vPOST(session,
                   path = paste0("folders/folders/"),
                   query = list(parentFolderUri = folders$items$parentFolderUri[1]),
                   payload = list(name = "newFolder"),
                   httr::content_type("application/json"))


deletedFolder <- vDELETE(session,
                         path = "folders/folders/",
                         resourceID = newFolder$id)

}
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}, \code{\link[httr:PUT]{httr::PUT()}}, \code{\link[httr:DELETE]{httr::DELETE()}}

Other core API requests:
\code{\link[=vDELETE]{vDELETE()}},
\code{\link[=vGET]{vGET()}},
\code{\link[=vHEAD]{vHEAD()}},
\code{\link[=vPUT]{vPUT()}}
}
\concept{core API requests}
