% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{write_in_out_json}
\alias{write_in_out_json}
\title{Write variable json}
\usage{
write_in_out_json(data, input = TRUE, path = NULL, noFile = FALSE)
}
\arguments{
\item{data}{\code{data.frame} to map the correct variable types to Json}

\item{input}{\code{TRUE} to write inputVar.json and \code{FALSE} to write \code{outputVar.json}}

\item{path}{directory where the JSON file is written; required when
\code{noFile = FALSE}}

\item{noFile}{if you don't want to write to a file, only the output}
}
\value{
\itemize{
\item \code{list} of the mapped types and sizes.
\item 'inputVar.json' or 'outputVar.json' file written to \code{path}
}
}
\description{
Writes a variable descriptor JSON file for input or output variables, based
on an input dataframe containing predictor and prediction columns.
}
\examples{

write_in_out_json(iris[,5, drop = FALSE], input = FALSE, noFile = TRUE)
write_in_out_json(iris[,1:4], input = TRUE, noFile = TRUE)

}
