% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{calculateLiftStat}
\alias{calculateLiftStat}
\title{Write dmcas_lift Json}
\usage{
calculateLiftStat(
  validadedf,
  traindf,
  testdf,
  targetName,
  targetPredicted,
  targetEventValue = 1,
  path = "./",
  noFile = FALSE
)
}
\arguments{
\item{validadedf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{traindf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{testdf}{\code{data.frame} where the first column in the yActual (labels/value) and the second is yPrediction (target probability)}

\item{targetName}{target variable column name (actuals)}

\item{targetPredicted}{target variable predicted probability column name}

\item{targetEventValue}{target class name for ROC reference, if model is nominal, all other class will be counted as "not target"}

\item{path}{default to current work dir}

\item{noFile}{if you don't want to write to a file, only the output}
}
\value{
\itemize{
\item \code{list} that reflects the 'dmcas_roc.json'
\item 'dmcas_roc.json' file written to \code{path}
}
}
\description{
Calculates the lift curves from user data and writes to a JSON file for
importing into the common model repository. Binary response only.
}
\examples{

df <- data.frame(label = sample(c(1,0), 6000, replace = TRUE),
                 prob = runif(6000),
                 partition = rep_len(1:3, 6000)) ## partition will be ignored since it is 3rd column
               
calculateLiftStat(df[df$partition == 1, ], 
                   df[df$partition == 2, ],
                   df[df$partition == 3, ],
                   targetName = "label",
                   targetPredicted = "prob",
                   noFile = TRUE)
                                    
                                    
 
}
