% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen}
\alias{codegen}
\alias{codegen.lm}
\alias{codegen.glm}
\title{SAS standard score code generation Generic Function \strong{(EXPERIMENTAL)}}
\usage{
codegen(model, path, rds, ...)

\method{codegen}{lm}(model, path = "scoreCode.R", rds = "model.rds")

\method{codegen}{glm}(model, path = "scoreCode.R", rds = "model.rds", cutoff = 0.5)
}
\arguments{
\item{model}{model object (lm, glm ...)}

\item{path}{file name and path to write}

\item{rds}{.rds file name to be called}

\item{...}{to be passes to individual code generators}
}
\value{
a code string
}
\description{
\strong{EXPERIMENTAL STATE - MAY NO WORK AS INTENDED}

Score code will only be generated successfully for supported models.
Other models and frameworks will be added in due time.
Use \code{\link[=create_scoreSample]{create_scoreSample()}} to get a structure sample
}
\section{Methods (by class)}{
\itemize{
\item \code{codegen(lm)}: Code generator for Linear models

\item \code{codegen(glm)}: generator for \code{glm} General Linear models, specifically logistic regression

}}
\examples{

\dontrun{
# SAS viya doesn't play nice with variables with '.' in the names
colnames(iris) <- gsub("\\\\.", "_", colnames(iris))

# simple regression
model <- lm(Petal.Length ~ ., data = iris)

codegen(model)
} 

}
