% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen}
\alias{codegen}
\alias{codegen.lm}
\alias{codegen.glm}
\alias{codegen.workflow}
\title{SAS standard score code generation Generic Function \strong{(EXPERIMENTAL)}}
\usage{
codegen(model, path, rds, libs, inputs, ...)

\method{codegen}{lm}(model, path = "scoreCode.R", rds = "model.rds", libs = c())

\method{codegen}{glm}(
  model,
  path = "scoreCode.R",
  rds = "model.rds",
  cutoff = 0.5,
  libs = c()
)

\method{codegen}{workflow}(
  tm_workflow,
  path = "scoreCode.R",
  rds = "model.rds",
  inputs = NULL,
  libs = c(),
  referenceLevel = NULL
)
}
\arguments{
\item{model}{model object (lm, glm ...)}

\item{path}{file name and path to write}

\item{rds}{.rds file name to be called}

\item{libs}{vector of libraries to be added to the code. Some may be guessed from the type.}

\item{inputs}{define inputs as the passed vector instead of guessed}

\item{...}{to be passes to individual code generators}
}
\value{
a code string
}
\description{
\strong{EXPERIMENTAL STATE - MAY NO WORK AS INTENDED}

Score code will only be generated successfully for supported models.
Other models and frameworks will be added in due time.
Use \code{\link[=create_scoreSample]{create_scoreSample()}} to get a structure sample
}
\section{Methods (by class)}{
\itemize{
\item \code{codegen(lm)}: Code generator for \code{lm} class models

\item \code{codegen(glm)}: generator for \code{glm} class models, specifically logistic regression

\item \code{codegen(workflow)}: generator for \link{tidymodels} \code{workflow} class models

}}
\examples{

\dontrun{
# SAS viya doesn't play nice with variables with '.' in the names
colnames(iris) <- gsub("\\\\.", "_", colnames(iris))

# simple regression
model <- lm(Petal.Length ~ ., data = iris)

codegen(model)
} 

}
