% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen}
\alias{codegen}
\title{SAS standard score code generation \strong{(EXPERIMENTAL)}}
\usage{
codegen(model, path, rds, ...)
}
\arguments{
\item{model}{model object (lm, glm ...)}

\item{path}{file name and path to write}

\item{rds}{.rds file name to be called}

\item{...}{to be passes to individual code generators}
}
\value{
a code string
}
\description{
\strong{EXPERIMENTAL STATE - MAY NO WORK AS INTENDED}

Score code will only be generated successfully for supported models.
Other models and frameworks will be added in due time.
Use \code{\link[=create_scoreSample]{create_scoreSample()}} to get a structure sample
}
\examples{

# SAS viya doesn't play nice with variables with '.' in the names
iris <- gsub("\\\\.", "_", colnames(iris))

# simple regression
model <- lm(Petal.Length ~ ., data = iris)

codegen(model)
 

}
