% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microanalyticScore.R
\name{delete_masmodule}
\alias{delete_masmodule}
\title{Delete a module/model and steps}
\usage{
delete_masmodule(session, module, exact = TRUE)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{module}{\code{MASmodule} object, module/model name. If name, will try to find a single module with exact name match. See \code{exact} parameter}

\item{exact}{the filter query should use "contains" for partial match or "eq" for exact match}
}
\value{
A \code{\link[httr:response]{httr::response}} object.
}
\description{
Delete a module/model published on MAS.
}
\examples{

\dontrun{
deleted_module <- delete_masmodule(sess, module = "ModuleName")
deleted_module
}

}
