% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{oauth_consul}
\alias{oauth_consul}
\title{Viya oauth token}
\usage{
oauth_consul(hostname, consul_token, verbose = FALSE)
}
\arguments{
\item{hostname}{string, SAS Viya url}

\item{consul_token}{consul token uuid}

\item{verbose}{logical, return print API call information}
}
\value{
list of API call request data
}
\description{
Requests a viya oauthtoken
}
\examples{

\dontrun{
token <- oauth_consul("http://myserver.com",
                  consul_token = "47817a5a-3751-4fad-9558-b12b8a702b69") #token is an uuid

}

}
\seealso{
Other authentication: 
\code{\link{session}()}
}
\concept{authentication}
