% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{vGET}
\alias{vGET}
\title{Viya GET}
\usage{
vGET(session, path, ..., query, verbose = FALSE, output = "json")
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{path}{character, path to the GET api endpoint}

\item{...}{additional parameters to be passed to \code{httr::GET} such as \code{httr::add_headers}}

\item{query}{list, additional URL query parameters}

\item{verbose}{logical, return print API call information}

\item{output}{string, if \code{output = "json"} will return \code{httr::fromJSON(httr::content(response, as = "text"))},
else if \code{output = "response"} will return a \code{httr::response()} object}
}
\value{
list if \code{output = "json"} default. \code{httr::response()} if \code{output = response}.
}
\description{
Wrapper to make generic GET calls to SAS Viya
}
\examples{

\dontrun{
folders <- vGET(session,
                path = "folders/folders/")


newFolder <- vPOST(session,
                   path = paste0("folders/folders/"),
                   query = list(parentFolderUri = folders$items$parentFolderUri[1]),
                   payload = list(name = "newFolder"),
                   httr::content_type("application/json"))


deletedFolder <- vDELETE(session,
                         path = "folders/folders/",
                         resourceID = newFolder$id)

}
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}, \code{\link[httr:PUT]{httr::PUT()}}, \code{\link[httr:DELETE]{httr::DELETE()}}

Other core API requests: 
\code{\link{vDELETE}()},
\code{\link{vHEAD}()},
\code{\link{vPOST}()},
\code{\link{vPUT}()}
}
\concept{core API requests}
