% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{vPUT}
\alias{vPUT}
\title{Viya PUT}
\usage{
vPUT(
  session,
  path,
  payload,
  ...,
  verbose = FALSE,
  encode = "json",
  output = "json"
)
}
\arguments{
\item{session}{viya_connection object, obtained through \code{session} function}

\item{path}{character, path to the GET api endpoint}

\item{payload}{list or json string, if it is a list, will be transformed in a json string using \code{jsonlite::toJSON}}

\item{...}{additional parameters to be passed to \code{httr::PUT} such as \code{httr::add_headers}}

\item{verbose}{logical, return print API call information}

\item{encode}{payload encoding type, to be passed to \code{httr::POST}.}

\item{output}{string, if \code{output = "json"} will return \code{httr::fromJSON(httr::content(response, as = "text"))},
else if \code{output = "response"} will return a \code{httr::response()} object}
}
\value{
list if \code{output = "json"} default. \code{httr::response()} if \code{output = response}.
}
\description{
Wrapper to make generic DELETE calls to SAS Viya
}
\details{
This function in built on top of httr for convenience when calling SAS Viya API endpoints..
}
\examples{

\dontrun{
folders <- vGET(session,
                path = "folders/folders/")


newFolder <- vPOST(session,
                   path = paste0("folders/folders/"),
                   query = list(parentFolderUri = folders$items$parentFolderUri[1]),
                   payload = list(name = "newFolder"),
                   httr::content_type("application/json"))


deletedFolder <- vDELETE(session,
                         path = "folders/folders/",
                         resourceID = newFolder$id)

}
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}, \code{\link[httr:PUT]{httr::PUT()}}, \code{\link[httr:DELETE]{httr::DELETE()}}, \code{\link[httr:response]{httr::response()}}

Other core API requests: 
\code{\link{vDELETE}()},
\code{\link{vGET}()},
\code{\link{vHEAD}()},
\code{\link{vPOST}()}
}
\concept{core API requests}
