% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{write_fileMetadata_json}
\alias{write_fileMetadata_json}
\title{Write fileMetadata json}
\usage{
write_fileMetadata_json(
  scoreCodeName = "scoreCode.R",
  scoreResource = "model.rda",
  additionalFilesNames = c(),
  additionalFilesRoles = c(),
  path = "./",
  noFile = FALSE
)
}
\arguments{
\item{scoreCodeName}{Name of the scoring code file}

\item{scoreResource}{rda file name or other score resources.}

\item{additionalFilesNames}{additional files names.}

\item{additionalFilesRoles}{additional files role names.}

\item{path}{filepath where to write the json (don't include the filename)}

\item{noFile}{if you don't want to write to a file, only list the output}
}
\value{
\itemize{
\item \code{list} of the mapped properties and values.
\item 'ModelProperties.json' file written to \code{path}
}
}
\description{
Writes a variable descriptor JSON file for fileMetadata, it will configure
the models files metadata within Model Manager in the first upload
}
\examples{

## Using default names and files
write_fileMetadata_json(noFile = TRUE)

## addition file resources
## send 2 vectors with file names and role name, must be of same length

write_fileMetadata_json(additionalFilesNames = c("myFileName.ext", "myFileName2.ext"),
                        additionalFilesRoles = c("scoreResource", "scoreResource"),
                        noFile = TRUE
                        )


}
