% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_files.R
\name{write_ModelProperties_json}
\alias{write_ModelProperties_json}
\title{Write ModelProperties json}
\usage{
write_ModelProperties_json(
  modelName,
  modelDescription = "R model",
  modelFunction,
  trainTable = " ",
  algorithm,
  numTargetCategories,
  targetEvent,
  targetVariable,
  eventProbVar,
  modeler = " ",
  tool = "R",
  toolVersion = "default",
  path = "./",
  noFile = FALSE
)
}
\arguments{
\item{modelName}{Name of the model}

\item{modelDescription}{String describing the model}

\item{modelFunction}{Classification, Prediction, Segmentation, Analytical or Clustering.}

\item{trainTable}{Name of the training table}

\item{algorithm}{Algorithm name (Random Forest, GLM, Linear Regression, etc.)}

\item{numTargetCategories}{number of possible classes for classification}

\item{targetEvent}{Target event label eg: "1", "versicolor" etc.}

\item{targetVariable}{Target variable name}

\item{eventProbVar}{Variable name that has the \code{targetEvent} probability Chance}

\item{modeler}{Modeler's name}

\item{tool}{Name of the tool used to build the model}

\item{toolVersion}{Version of the tool used to build the model}

\item{path}{file path where to write the json (don't include the filename)}

\item{noFile}{if you don't want to write to a file, only list the output}
}
\value{
\itemize{
\item \code{list} of the mapped properties and values.
\item 'ModelProperties.json' file written to \code{path}
}
}
\description{
Writes a descriptor JSON file for ModelProperties, it will configure
the model properties within Model Manager
}
\examples{

write_ModelProperties_json(modelName = "My R Model", 
                           modelDescription = "Awesome Description", 
                           modelFunction = "Classification",
                           trainTable = " ",
                           algorithm = "Logistic Regression",
                           numTargetCategories = 2,
                           targetEvent = "BAD",
                           targetVariable = "P_BAD1",
                           eventProbVar = "P_BAD1",
                           modeler = "John SAS",
                           noFile = TRUE)

}
